Extract functions to child module
